+++ setup.sh
@@ -1,0 +1,1 @@
+../setup.sh
